setup plugin and base on sbt-paradox-material-theme
