+++ build.sbt
@@ -39,7 +39,8 @@ lazy val pekkoTheme = project
   .enablePlugins(ParadoxThemePlugin)
   .settings(
     organization := "org.apache.pekko",
-    name := "pekko-theme-paradox")
+    name := "pekko-theme-paradox",
+    libraryDependencies += "io.github.jonas" % "paradox-material-theme" % "0.6.0")
   .settings(publishSettings)
 
 lazy val pekkoPlugin = project
@@ -57,9 +58,10 @@ lazy val pekkoPlugin = project
     addSbtPlugin(
       // When updating the sbt-paradox version,
       // remember to also update project/plugins.sbt
-      "com.lightbend.paradox" % "sbt-paradox" % "0.10.3"),
-    addSbtPlugin("com.lightbend.paradox" % "sbt-paradox-apidoc" % "1.0.0"),
-    addSbtPlugin("com.lightbend.paradox" % "sbt-paradox-project-info" % "3.0.0"),
+      "com.lightbend.paradox" % "sbt-paradox" % "0.9.2"),
+    addSbtPlugin("com.lightbend.paradox" % "sbt-paradox-apidoc" % "0.10.1"),
+    addSbtPlugin("com.lightbend.paradox" % "sbt-paradox-project-info" % "2.0.0"),
+    addSbtPlugin("io.github.jonas" % "sbt-paradox-material-theme" % "0.6.0"),
     Compile / resourceGenerators += Def.task {
       val file = (Compile / resourceManaged).value / "pekko-paradox.properties"
       IO.write(file, s"pekko.paradox.version=${version.value}")
